also move ghutil and slackutil
