+++ test/mako/alerter/github/issue_test.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 	"testing"
 
-	"knative.dev/test-infra/shared/ghutil"
-	"knative.dev/test-infra/shared/ghutil/fakeghutil"
+	"knative.dev/pkg/test/ghutil"
+	"knative.dev/pkg/test/ghutil/fakeghutil"
 )
 
 var gih issueHandler
